Migrated the thread pool structures to thread_pool_internal.h.
Also created a couple accessor functions for thread pool size
(no. threads), thread queue size, and extracting the void* data from a
thread "result".

[Updated Makefile dependencies -- JM]
