+++ Makefile
@@ -188,6 +188,7 @@ cram_structs_h = cram/cram_structs.h htslib/thread_pool.h cram/string_alloc.h $(
 cram_open_trace_file_h = cram/open_trace_file.h cram/mFILE.h
 hfile_internal_h = hfile_internal.h $(htslib_hfile_h)
 hts_internal_h = hts_internal.h $(htslib_hts_h)
+thread_pool_internal_h = thread_pool_internal.h htslib/thread_pool.h
 
 
 # To be effective, config.mk needs to appear after most Makefile variables are
@@ -303,7 +304,7 @@ cram/pooled_alloc.o cram/pooled_alloc.pico: cram/pooled_alloc.c config.h cram/po
 cram/rANS_static.o cram/rANS_static.pico: cram/rANS_static.c config.h cram/rANS_static.h cram/rANS_byte.h
 cram/sam_header.o cram/sam_header.pico: cram/sam_header.c config.h $(cram_sam_header_h) cram/string_alloc.h
 cram/string_alloc.o cram/string_alloc.pico: cram/string_alloc.c config.h cram/string_alloc.h
-thread_pool.o thread_pool.pico: thread_pool.c config.h htslib/thread_pool.h
+thread_pool.o thread_pool.pico: thread_pool.c config.h $(thread_pool_internal_h)
 cram/vlen.o cram/vlen.pico: cram/vlen.c config.h cram/vlen.h cram/os.h
 cram/zfio.o cram/zfio.pico: cram/zfio.c config.h cram/os.h cram/zfio.h
